Home-PR-1: module launcher under Hero — shared CreateTripForm, live Travel (register-gated), 5 paid stubs
